Use visible/gone instead of scaling X/Y on old API for label text.
Note this could also make use of ViewCompat, but decided not to.
The reason is that ViewCompat contains too many no-ops and it will
hide the fact that the UI is not actually doing what is required.
Calling ViewCompat.setScaleX will do nothing but the compiller will
not know that the UI will be broken after executing that (because
it did nothing).

Also note that the `start()` animation method is not required. The
animation will start on the next frame anyway.

Finally, needed to move the call to `super.select()` and friends
to the end of the method in the subclass. This is because the subclass
changes the visibility of the text view. This visibility is then queried
by the base class so it needs to be set before invoking `super.*()`.